Refactor the card grid from Layout into CardGrid component.
